# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expoza API",
        "description": "Expoza API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/send-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send OTP to user (signup or forgot password)",
                "description": "Step 1: User enters mobile, receives OTP via SMS. Use this for both signup and forgot password.",
                "operationId": "5cc0f5fdd1bd643c4735a0535615dafc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "purpose"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "enum": [
                                            "signup",
                                            "forgot"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "otp": {
                                            "type": "string",
                                            "example": "123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify OTP for user (signup or forgot password)",
                "description": "Step 2: User enters mobile and OTP. If OTP is valid, returns a temp_token (required for signup).",
                "operationId": "59582211905db382e25251f34da30599",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "otp"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified successfully. Returns temp_token for next step.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "temp_token": {
                                            "type": "string",
                                            "example": "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired OTP"
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Signup with mobile, temp_token, name, password (after OTP verified)",
                "description": "Step 3: User enters mobile, temp_token (from verify-otp), name, password, and optional email. Completes signup and receives login token.",
                "operationId": "168042e7139237a6632fce73be675384",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "temp_token",
                                    "full_name",
                                    "password"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "temp_token": {
                                        "type": "string",
                                        "example": "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Signup successful. Returns user and login token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired signup token"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}